Rename Terminal Type for Terminfo Database
I set up the terminal name to a custom value based on some HOWTOs
I read when originally setting this up. This custom version supports
italic text. The custom name caused some compatibility issues with
other software and systems. I found that the new terminfo entry is
stored just under my account, and not the system as a whole, so it
is fine to use the common name. So I did.