Zsh: Add Multi-Architecture Support
Check which archicture the terminal is running under and configure paths
for some software to use i386 binaries instead of arm64 binaries. I set
this up because I found that the Oracle Instantclient libraries aren't
available for Apple's arm64 architecture so I need to run scripts that
depend on them in a terminal running under Rosetta.