Updates to bash startup/shutdown configuration
Add vim modelines to set filetypes to "bash" so the editor uses proper
syntax highlighting and linting/formatting.

Don't start TMUX automatically when logging in with bash shell. Since we
use Zsh for most work trim down the bash start scripts so we can use
that for access if tmux or zsh gets in a weird state.