Reorganize $PATH setup
I dropped the ~/.zprofile in favor of ~/.zshenv because it's the first
file read on zsh startup. I also moved PATH setup to an oh-my-zsh custom
setup that gets run after all plugins are initialized to get better
control of how they are ordered after the path is cleaned up at the end
of .zshrc.